Added Team Role Models and Default Data.
Altered Production Config to define SQLAlchemy URI to get rid
of a warning.
Changed all get views to use an envelope called 'data' which will hold
the resources.
Added data upgrade/downgrade for default role types.
Changed all tests to use the migrations for the db creation/deletion,
so that it is closer to the real system and will get default data.