Added Mood Template Table
I also addressed a lot of issues with marshalling, and
generally how we handle the different responses for get and
post. We now expect that we should either have a success such
as 200/201 and the resourse enclosed in a 'data' field. Or we
have a failure that has a status and message value.

Also aligned a lot of the way the view functions are laid out.
Finally I fixed a few defects and added tests to cover them.
Things such as the user view outputting the password object.